+++ bookeeper/src/main/java/com/etn319/service/impl/BookServiceImpl.java
@@ -1,15 +1,15 @@
 package com.etn319.service.impl;
 
-import com.etn319.dao.DaoLayerException;
-import com.etn319.dao.api.AuthorDao;
-import com.etn319.dao.api.BookDao;
-import com.etn319.dao.api.GenreDao;
+import com.etn319.dao.datajpa.AuthorRepository;
+import com.etn319.dao.datajpa.BookRepository;
+import com.etn319.dao.datajpa.GenreRepository;
 import com.etn319.model.Book;
 import com.etn319.service.CacheHolder;
 import com.etn319.service.ServiceLayerException;
 import com.etn319.service.api.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,9 +21,9 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Service
 public class BookServiceImpl implements BookService {
-    private final BookDao dao;
-    private final AuthorDao authorDao;
-    private final GenreDao genreDao;
+    private final BookRepository dao;
+    private final AuthorRepository authorDao;
+    private final GenreRepository genreDao;
     private final CacheHolder cache;
 
     @Override
@@ -51,7 +51,7 @@ public class BookServiceImpl implements BookService {
             Book saved = dao.save(book);
             clearCache();
             return saved;
-        } catch (DaoLayerException e) {
+        } catch (DataAccessException e) {
             throw new ServiceLayerException(e);
         }
     }
@@ -59,10 +59,15 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional
     public void deleteById(long id) {
-        try {
-            dao.deleteById(id);
-        } catch (DaoLayerException e) {
-            throw new ServiceLayerException(e);
+        if (dao.existsById(id)) {
+            try {
+                dao.deleteById(id);
+            } catch (DataAccessException e) {
+                throw new ServiceLayerException(e);
+            }
+        } else {
+            // todo: заполнить исключение или придумать новое
+            throw new ServiceLayerException();
         }
     }
