+++ bookeeper/src/main/java/com/etn319/service/impl/GenreServiceImpl.java
@@ -1,13 +1,13 @@
 package com.etn319.service.impl;
 
-import com.etn319.dao.DaoLayerException;
-import com.etn319.dao.api.GenreDao;
+import com.etn319.dao.datajpa.GenreRepository;
 import com.etn319.model.Genre;
 import com.etn319.service.CacheHolder;
 import com.etn319.service.ServiceLayerException;
 import com.etn319.service.api.GenreService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +18,7 @@ import java.util.Optional;
 @Slf4j
 @RequiredArgsConstructor
 public class GenreServiceImpl implements GenreService {
-    private final GenreDao dao;
+    private final GenreRepository dao;
     private final CacheHolder cache;
 
     @Override
@@ -46,7 +46,7 @@ public class GenreServiceImpl implements GenreService {
             Genre saved = dao.save(genre);
             clearCache();
             return saved;
-        } catch (DaoLayerException e) {
+        } catch (DataAccessException e) {
             throw new ServiceLayerException(e);
         }
     }
@@ -54,10 +54,15 @@ public class GenreServiceImpl implements GenreService {
     @Override
     @Transactional
     public void deleteById(long id) {
-        try {
-            dao.deleteById(id);
-        } catch (DaoLayerException e) {
-            throw new ServiceLayerException(e);
+        if (dao.existsById(id)) {
+            try {
+                dao.deleteById(id);
+            } catch (DataAccessException e) {
+                throw new ServiceLayerException(e);
+            }
+        } else {
+            // todo: заполнить исключение или придумать новое
+            throw new ServiceLayerException();
         }
     }
